refactor(sensors): unify three-write diagnostics and auto-drive lifecycle recorders
- 三写约束：所有中止路径写入 abort_reason（粗枚举）+ DiagnosticIssueList + PROJECT_LOG
- SAR abort_reason 从 39 个字符串瘦身为 6 个粗粒度枚举，细粒度细节进 diagnostics
- 路线 A：四个 Session（AR/EOS/SAR/SBIRS）可选注册 LifecycleRecorder，
  StepWithResult 内部自动驱动 Update，消除调用方漏调失步陷阱
- Recorder 新增 GetLastEvents() 缓存最近一次执行周期的事件
- kProductUpdated 改名 kProductSustained（如实表达"产品持续存在"）
- SBIRS powered-off 迁移收尾：replay codec 白名单 + batch_validation 期望值 + roundtrip 测试
- 文档：session_contract.md 规则 10-11（两种驱动方式 + Attach 契约）

+++ examples/sar/session_usage.cpp
@@ -94,7 +94,7 @@ int main() {
   const sar::session::SarCycleResult result = session.StepWithResult(MakeInput());
 
   if (result.has_error) {
-    std::cerr << "SAR processing failed: " << result.abort_reason << '\n';
+    std::cerr << "SAR processing failed: " << static_cast<int>(result.abort_reason) << '\n';
     for (const sar::session::SarDiagnosticIssue& issue : result.diagnostics) {
       std::cerr << '[' << SeverityName(issue.severity) << "] " << issue.code << ": "
                 << issue.message << '\n';
